pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drf-jsonmask"
dynamic = ["version"]
description = "Implements Google's partial response in Django RestFramework (Fork from Zapier's package)"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    { name = "Craig Labenz", email = "craig.labenz@zapier.com" },
    { name = "João Luiz Lorencetti", email = "jll.linux@gmail.com" },
]
maintainers = [
    { name = "Christian Hartung", email = "hartungstenio@outlook.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "jsonmask",
    "django>=2.2,<4.1",
    "djangorestframework>=3.10,<3.14",
]

[project.urls]
Repository = "https://github.com/jllorencetti/drf-jsonmask.git"

[tool.hatch.version]
path = "drf_jsonmask/__init__.py"

[tool.coverage.run]
source = ["drf_jsonmask"]
branch = true

[tool.isort]
line_length = 119
combine_as_imports = true
include_trailing_comma = true
multi_line_output = 5
default_section = "THIRDPARTY"
